requirements.txt
helium
beautifulsoup4
langchain
langchain-community
langchain-huggingface
streamlit
python-dotenv
requests-html
google.generativeai